requirements.txt
torch>=1.10.0
torchvision>=0.11.0
Pillow>=8.4.0
numpy>=1.21.0
requests>=2.26.0
